# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Essensplan API
  description: |-
    
  contact:
    email: support@schulzentrum.de
  license:
    name: Open Source License 3.0
    url: https://opensource.org/license/osl-3-0
  version: 0.9.1
servers:
  - url: http://localhost
  - url: https://secure-gateway.schulzentrum.de/api/v1/essensplan
tags:
  - name: plans
    description: Daily and weekly meal plans
  - name: meals
    description: Information about meals
  - name: specifics
    description: More information on details like allergens and supplements
  - name: metadata
    description: Metadata information like available years
  - name: openmensa
    description: Endpoints that are compatibil with OpenMensa APIs
    externalDocs:
      description: Developer Documentation
      url: https://docs.openmensa.org/
paths:
  /allergens:
    get:
      tags:
        - specifics
      summary: Returns list of all known allergens
      responses:
        "200":
          description: List of all allergens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Allergen"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /allergens/{id}:
    parameters:
      - name: id
        description: ID of a specific allergen
        in: path
        required: true
        schema: 
          type: string
          enum: ["A", "B", "C"]
    get:
      tags:
        - specifics
      summary: Returns information on a specific allergen
      responses:
        "200":
          description: Allergen details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Allergen"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /supplements:
    get:
      tags:
        - specifics
      summary: Returns list of all known supplements
      responses:
        "200":
          description: List of all supplements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplement"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /supplements/{id}:
    parameters:
      - name: id
        description: ID of a specific supplement
        in: path
        required: true
        schema: 
          type: string
          enum: ["A", "B", "C"]
    get:
      tags:
        - specifics
      summary: Returns information on a specific supplement
      responses:
        "200":
          description: Supplement details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplement"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /menues:
    get:
      tags:
        - specifics
      summary: Returns list of all known menues/meal categories
      responses:
        "200":
          description: List of all menu categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /meals:
    get:
      tags:
        - meals
      summary: Query all known meals
      responses:
        "200":
          description: List of all meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /meals/{mealID}:
    parameters:
      - name: mealID
        description: ID of a specific meal
        in: path
        required: true
        schema: 
          type: integer
          format: int
    get:
      tags:
        - meals
      summary: Return information on a specific meal
      responses:
        "200":
          description: Meal details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /weeks/{date}:
    parameters:
      - name: date
        description: Date in requested week
        in: path
        required: true
        schema: 
          type: string
          format: date-time
    get:
      tags:
        - plans
      summary: Return weekly meal plan for a given date in that week
      description: Return weekly meal plan for a given date in that week
      responses:
        "200":
          description: Weekly meal plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DateRangePlan"
        "400":
          description: Invalid date format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /days/{date}:
    parameters:
      - name: date
        description: Date in requested week
        in: path
        required: true
        schema: 
          type: string
          format: date-time
    get:
      tags:
        - plans
      summary: Return daily meal plan for a given date
      description: Return daily meal plan for a given date
      responses:
        "200":
          description: Daily meal plan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyPlanEntry"
        "400":
          description: Invalid date format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /plan-range/{startDate}/{endDate}:
    parameters:
      - name: startDate
        description: Starting date for plan range
        in: path
        required: true
        schema: 
          type: string
          format: date-time
      - name: endDate
        description: Ending date for plan range
        in: path
        required: true
        schema: 
          type: string
          format: date-time
    get:
      tags:
        - plans
      summary: Return meal plan for a given date range
      description: Return meal plan for a given date range
      responses:
        "200":
          description: Meal plan for the date range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DateRangePlan"
        "400":
          description: Invalid date format, invalid date range, or range limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /plans:
    get:
      tags:
        - plans
      summary: Return meal plans for all days from today on that have a meal plan
      description: Return meal plans for all days from today on that have a meal plan
      responses:
        "200":
          description: Meal plans from today onwards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DateRangePlan"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /plans/{relativeDayOffset}:
    parameters:
      - name: relativeDayOffset
        description: Relative day offset from today
        in: path
        required: true
        schema: 
          type: integer
          format: int
          min: -999
          max: 999
    get:
      tags:
        - plans
      summary: Return meal plans for all days from today until day offset
      description: Return meal plans for all days from today until day offset including future and previous days which have a meal plan
      responses:
        "200":
          description: Meal plans for the offset range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DateRangePlan"
        "400":
          description: Invalid range limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /available-days/{year}:
    parameters:
      - name: year
        description: Year for which available dates are requested
        in: path
        required: true
        schema: 
          type: integer
          format: integer
    get:
      tags:
        - metadata
      summary: Returns list of dates for which meal plans are available in a given year
      responses:
        "200":
          description: List of available dates in the given year
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: date
        "500":
          $ref: "#/components/responses/InternalServerError"
  /years:
    get:
      tags:
        - metadata
      summary: Returns list of years for which meal plans are available
      responses:
        "200":
          description: List of available years
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
        "500":
          $ref: "#/components/responses/InternalServerError"
  /openmensa/feed-v2:
    get:
      tags: 
        - openmensa
      summary: OpenMensa parser feed endpoint
      externalDocs:
        description: Find out more about OpenMensa parser feed
        url: https://docs.openmensa.org/feed/v2/
      responses:
        "200":
          description: "OK"
          content:
            application/xml:
              schema:
  /openmensa/api-v2/canteens:
    get:
      tags:
        - openmensa
      summary: List all canteens
  /openmensa/api-v2/canteens/{canteen_id}:
    parameters:
      - name: canteen_id
        description: ID of canteen
        in: path
        required: true
        schema: 
          type: integer
          format: int
          enum: [1]
    get:
      tags:
        - openmensa
      summary: Get a canteen
  /openmensa/api-v2/canteens/{canteen_id}/days:
    parameters:
      - name: canteen_id
        description: ID of canteen
        in: path
        required: true
        schema: 
          type: integer
          format: int
          enum: [1]
    get:
      tags:
        - openmensa
      summary: List days of canteen
      responses: 
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/day"
  /openmensa/api-v2/canteens/{canteen_id}/days/{date}:
    parameters:
      - name: canteen_id
        description: ID of canteen
        in: path
        required: true
        schema: 
          type: integer
          format: int
          enum: [1]
      - name: date
        description: date of day
        in: path
        required: true
        schema: 
          type: string
          format: date
    get:
      tags:
        - openmensa
      summary: Get day of canteen
      responses: 
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/day"
  /openmensa/api-v2/canteens/{canteen_id}/days/{date}/meals:
    parameters:
      - name: canteen_id
        description: ID of canteen
        in: path
        required: true
        schema: 
          type: integer
          format: int
          enum: [1]
      - name: date
        description: date of day
        in: path
        required: true
        schema: 
          type: string
          format: date
    get:
      tags:
        - openmensa
      summary: List meals of a day in a canteen
      responses: 
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/meal"
  /openmensa/api-v2/canteens/{canteen_id}/days/{date}/meals/{id}:
    parameters:
      - name: canteen_id
        description: ID of canteen
        in: path
        required: true
        schema: 
          type: integer
          format: int
          enum: [1]
      - name: date
        description: date of day
        in: path
        required: true
        schema: 
          type: string
          format: date
      - name: id
        description: ID of meal
        in: path
        required: true
        schema: 
          type: integer
          format: int
    get:
      tags:
        - openmensa
      summary: Get a meal
      responses: 
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/meal"
components:
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        type:
          type: string
          enum:
            - invalid_date_format
            - invalid_date_range
            - range_limit_exceeded
      required:
        - error
    Allergen:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
    Supplement:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
    Meal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dietary:
          type: string
          description: Comma-separated dietary labels (vegetarian, vegan, gluten-free)
        allergens:
          type: array
          items:
            $ref: "#/components/schemas/Allergen"
        supplements:
          type: array
          items:
            $ref: "#/components/schemas/Supplement"
        created_date:
          type: string
          format: date-time
      required:
        - id
        - name
        - dietary
        - allergens
        - supplements
        - created_date
    DailyPlanEntry:
      type: object
      properties:
        category:
          $ref: "#/components/schemas/Category"
        meal:
          $ref: "#/components/schemas/Meal"
        displayOrder:
          type: integer
      required:
        - category
        - meal
        - displayOrder
    DateRangePlanEntry:
      type: object
      properties:
        date:
          type: string
          format: date-time
        category:
          $ref: "#/components/schemas/Category"
        meal:
          $ref: "#/components/schemas/Meal"
        displayOrder:
          type: integer
      required:
        - date
        - category
        - meal
        - displayOrder
    DateRangePlan:
      type: object
      additionalProperties:
        type: array
        items:
          $ref: "#/components/schemas/DateRangePlanEntry"
      description: Object with date strings as keys (YYYY-MM-DD format) and arrays of meal plan entries as values
    meal:
      type: object
      properties:
        id: 
          type: integer
        name: 
          type: string
        notes: 
          type: array
          minItems: 0
          items:
            type: string
        prices:
          description: >
            Prices for different groups, which will be all the same because
            we don't differentiate the groups.
          type: object
          properties:
            pupils:
              optional: true
              type: number
            students:
              optional: true
              type: number
            employees:
              optional: true
              type: number
            others:
              optional: true
              type: number
        category:
          type: string
    day:
      type: object
      properties:
        date:
          type: string
        closed:
          type: boolean
    canteen:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        address:
          type: string
        coordinates:
          type: array
          items:
            type: number
          minItems: 2
          maxItems: 2